kulbaev_azat hw 5/requirements.txt
python-decouple==3.6
termcolor==1.1.0
